mcal/functions/sample_app/requirements.txt
azure-functions
requests==2.31.0
boto3==1.35.0